Create text_scrapper.py
In this file, we are extracting text from a website and storing it in a comma separated value file(csv). In detail, we are exporting the libraries needed to achieve our goal. We need the selenium lib to scrape the data and csv to create that file and its contents. Where you find 'yourthing' this will refer to what element you want to scrape. i assume you know how to find it already( if not, just right click the element then chose inspect it, you should then copy the text after "<.. class=" Do not include the actual text you want to scrape.